Move shop finder above gear guide
