+++ src/components/assessment/BikeAssessment.vue
@@ -84,11 +84,6 @@
             @bike-change="handleBikeChange"
           />
 
-          <bike-shop-finder
-            v-if="recommendation"
-            :recommended-bike-type="recommendation"
-          />
-
           <nav class="results-jump-links" aria-label="Jump to section">
             <div class="jump-label">Quick Links</div>
 
@@ -118,10 +113,15 @@
             :all-bike-types="bikeTypeDetails"
             :selected-bike-type="recommendation"
             :profile="assessmentProfile"
-            @bike-change="handleBikeChange"
+            :comparison-bike="comparisonBike"
             @update:savings="stickysavings = $event"
             @update:car-label="stickyCarLabel = $event"
             key="savings-component"
+          />
+
+          <bike-shop-finder
+            v-if="recommendation"
+            :recommended-bike-type="recommendation"
           />
 
           <div class="gear-guide-cta">
@@ -132,6 +132,17 @@
               View Essential Gear Guide
             </router-link>
           </div>
+
+          <bike-type-comparison-grid
+            v-if="recommendation"
+            :bike-title="recommendationDetails.title"
+            :bike-image="recommendationDetails.image"
+            :costs="costs"
+            :selected-bike-type="recommendation"
+            :all-bike-types="bikeTypeDetails"
+            :comparison-bike="comparisonBike"
+            @bike-change="handleComparisonBikeChange"
+          />
 
           <results-footer @restart="restartAssessment" />
         </div>
@@ -150,6 +161,7 @@ import FitnessStep from './FitnessStep.vue';
 import BikeRecommendation from './BikeRecommendation.vue';
 import BikeShopFinder from './BikeShopFinder.vue';
 import SavingsComparison from './SavingsComparison.vue';
+import BikeTypeComparisonGrid from './BikeTypeComparisonGrid.vue';
 import ResultsFooter from './ResultsFooter.vue';
 import YourChoices from './YourChoices.vue';
 import { CarCosts } from '../../constants/vehicleCosts';
@@ -193,6 +205,7 @@ watch(transportationNeeds, () => {}, { deep: true });
 watch(geography, () => {}, { deep: true });
 const recommendation = ref<BikeTypeId | ''>('');
 const alternateBikeType = ref<BikeTypeId | null>(null);
+const comparisonBike = ref('');
 
 watch(recommendation, (bikeTypeId) => {
   if (bikeTypeId) {
@@ -406,6 +419,16 @@ function handleBikeChange(bikeType: BikeTypeId | null) {
 
   // Update URL with the new bike selection
   updateUrlWithRecommendation(bikeType);
+}
+
+// Reset comparison when the base recommendation changes (e.g. URL navigation)
+watch(recommendation, (newType, oldType) => {
+  if (oldType && newType !== oldType) comparisonBike.value = '';
+});
+
+function handleComparisonBikeChange(bikeType: BikeTypeId) {
+  comparisonBike.value = bikeType;
+  handleBikeChange(bikeType);
 }
 
 // Function to update URL with bike recommendation
